refactor(tvix/eval): cargo clippy (len() is usize)
Reviewed-on: https://cl.tvl.fyi/c/depot/+/9099
Autosubmit: flokli <flokli@flokli.de>
Tested-by: BuildkiteCI

+++ eval/src/builtins/mod.rs
@@ -871,7 +871,7 @@ mod pure_builtins {
         }
 
         let end = if len < 0 {
-            x.as_str().len() as usize
+            x.as_str().len()
         } else {
             cmp::min(beg + (len as usize), x.as_str().len())
         };
